raw version LOGGER aop 2
