+++ src/main/resources/log4j.properties
@@ -3,13 +3,35 @@
 #See the configuration Tomkat
 #Tomkat in 9.0.0. Log files are written to the directory of apache-tomcat-9.0.0.M1/bin/
 
-log4j.rootLogger=INFO, infoLog
-#log4j.rootLogger=debugLog, infoLog, warnLog, errorLog, ALL
-
-#log4j.logger.debugLog=DEBUG, debugLog
-#log4j.logger.infoLog=INFO, infoLog
-#log4j.logger.warnLog=WARN, warnLog
-#log4j.logger.errorLog=ERROR, errorLog
+#log4j.rootLogger=OFF
+#
+#log4j.appender.debugLog=org.apache.log4j.DailyRollingFileAppender
+#log4j.appender.debugLog.File=../logs/debug.log
+#log4j.appender.debugLog.MaxFileSize=5MB
+#log4j.appender.debugLog.MaxBackupIndex=10
+#log4j.appender.debugLog.layout=org.apache.log4j.PatternLayout
+#log4j.appender.debugLog.layout.ConversionPattern=%d{yyyy-MM-dd HH:mm:ss} %-5p %c{1}:%L - %m%n
+#
+#log4j.appender.infoLog=org.apache.log4j.DailyRollingFileAppender
+#log4j.appender.infoLog.File=../logs/info.log
+#log4j.appender.infoLog.MaxFileSize=5MB
+#log4j.appender.infoLog.MaxBackupIndex=10
+#log4j.appender.infoLog.layout=org.apache.log4j.PatternLayout
+#log4j.appender.infoLog.layout.ConversionPattern=%d{yyyy-MM-dd HH:mm:ss} %-5p %c{1}:%L - %m%n
+#
+#log4j.appender.warnLog=org.apache.log4j.DailyRollingFileAppender
+#log4j.appender.warnLog.File=../logs/warn.log
+#log4j.appender.warnLog.MaxFileSize=5MB
+#log4j.appender.warnLog.MaxBackupIndex=10
+#log4j.appender.warnLog.layout=org.apache.log4j.PatternLayout
+#log4j.appender.warnLog.layout.ConversionPattern=%d{yyyy-MM-dd HH:mm:ss} %-5p %c{1}:%L - %m%n
+#
+#log4j.appender.errorLog=org.apache.log4j.DailyRollingFileAppender
+#log4j.appender.errorLog.File=../logs/error.log
+#log4j.appender.errorLog.MaxFileSize=5MB
+#log4j.appender.errorLog.MaxBackupIndex=10
+#log4j.appender.errorLog.layout=org.apache.log4j.PatternLayout
+#log4j.appender.errorLog.layout.ConversionPattern=%d{yyyy-MM-dd HH:mm:ss} %-5p %c{1}:%L - %m%n
 
 #log4j.category.debugLogger=DEBUG, debugLog
 #log4j.additivity.debugLogger=false
@@ -23,45 +45,3 @@ log4j.rootLogger=INFO, infoLog
 #log4j.category.errorLogger=ERROR, errorLog
 #log4j.additivity.errorLogger=false
 
-#log4j.appender.debugLog=org.apache.log4j.FileAppender
-##log4j.appender.debugLog.Threshold=DEBUG
-#log4j.appender.debugLog.File=../logs/debug.log
-#log4j.appender.debugLog.MaxFileSize=5MB
-#log4j.appender.debugLog.MaxBackupIndex=10
-#log4j.appender.debugLog.layout=org.apache.log4j.PatternLayout
-#log4j.appender.debugLog.layout.ConversionPattern=%d{yyyy-MM-dd HH:mm:ss} %-5p %c{1}:%L - %m%n
-
-log4j.appender.infoLog=org.apache.log4j.FileAppender
-log4j.appender.infoLog.Threshold=INFO
-log4j.appender.infoLog.File=../logs/info.log
-log4j.appender.infoLog.MaxFileSize=5MB
-log4j.appender.infoLog.MaxBackupIndex=10
-log4j.appender.infoLog.layout=org.apache.log4j.PatternLayout
-log4j.appender.infoLog.layout.ConversionPattern=%d{yyyy-MM-dd HH:mm:ss} %-5p %c{1}:%L - %m%n
-
-#log4j.appender.warnLog=org.apache.log4j.FileAppender
-##log4j.appender.warnLog.Threshold=WARN
-#log4j.appender.warnLog.File=../logs/warn.log
-#log4j.appender.warnLog.MaxFileSize=5MB
-#log4j.appender.warnLog.MaxBackupIndex=10
-#log4j.appender.warnLog.layout=org.apache.log4j.PatternLayout
-#log4j.appender.warnLog.layout.ConversionPattern=%d{yyyy-MM-dd HH:mm:ss} %-5p %c{1}:%L - %m%n
-#
-#log4j.appender.errorLog=org.apache.log4j.FileAppender
-##log4j.appender.errorLog.Threshold=ERROR
-#log4j.appender.errorLog.File=../logs/error.log
-#log4j.appender.errorLog.MaxFileSize=5MB
-#log4j.appender.errorLog.MaxBackupIndex=10
-#log4j.appender.errorLog.layout=org.apache.log4j.PatternLayout
-#log4j.appender.errorLog.layout.ConversionPattern=%d{yyyy-MM-dd HH:mm:ss} %-5p %c{1}:%L - %m%n
-
-#log4j.logger.evg.testt.controller.DepartmentController
-
-
-## Direct log messages to stdout
-#log4j.appender.stdout=org.apache.log4j.ConsoleAppender
-#log4j.appender.stdout.target=System.out
-#log4j.appender.stdout.layout=org.apache.log4j.PatternLayout
-#log4j.appender.stdout.layout.conversionPattern=%d{dd.MM.yy HH:mm:ss} %5p - %m%n
-#log4j.appender.stdout.encoding=UTF-8
-##
