+++ src/main/resources/log4j.xml
@@ -1,0 +1,62 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE log4j:configuration SYSTEM "log4j.dtd">
+
+<log4j:configuration debug="false" xmlns:log4j="http://jakarta.apache.org/log4j">
+
+    <appender name="debugLog" class="org.apache.log4j.DailyRollingFileAppender">
+        <param name="File" value="../logs/debug.log"/>
+        <param name="MaxFileSize" value="5MB"/>
+        <param name="MaxBackupIndex" value="10"/>
+        <layout class="org.apache.log4j.PatternLayout">
+            <param name="ConversionPattern" value="%d{yyyy-MM-dd HH:mm:ss} %-5p %c{1}:%L - %m%nn" />
+        </layout>
+    </appender>
+    
+    <appender name="infoLog" class="org.apache.log4j.DailyRollingFileAppender">
+        <param name="File" value="../logs/info.log"/>
+        <param name="MaxFileSize" value="5MB"/>
+        <param name="MaxBackupIndex" value="10"/>
+        <layout class="org.apache.log4j.PatternLayout">
+            <param name="ConversionPattern" value="%d{yyyy-MM-dd HH:mm:ss} %-5p %c{1}:%L - %m%nn" />
+        </layout>
+    </appender>
+    
+    <appender name="warnLog" class="org.apache.log4j.DailyRollingFileAppender">
+        <param name="File" value="../logs/warn.log"/>
+        <param name="MaxFileSize" value="5MB"/>
+        <param name="MaxBackupIndex" value="10"/>
+        <layout class="org.apache.log4j.PatternLayout">
+            <param name="ConversionPattern" value="%d{yyyy-MM-dd HH:mm:ss} %-5p %c{1}:%L - %m%nn" />
+        </layout>
+    </appender>
+    
+    <appender name="errorLog" class="org.apache.log4j.DailyRollingFileAppender">
+        <param name="File" value="../logs/error.log"/>
+        <param name="MaxFileSize" value="5MB"/>
+        <param name="MaxBackupIndex" value="10"/>
+        <layout class="org.apache.log4j.PatternLayout">
+            <param name="ConversionPattern" value="%d{yyyy-MM-dd HH:mm:ss} %-5p %c{1}:%L - %m%nn" />
+        </layout>
+    </appender>
+
+    <root>
+        <priority value="error"/>
+        <appender-ref ref="errorLog" />
+    </root>
+    
+    <logger name="org.hibernate" additivity="false">
+        <level value="warn"/>
+        <appender-ref ref="warnLog"/>
+    </logger>
+
+    <logger name="evg.testt">
+        <level value="info"/>
+        <appender-ref ref="infoLog"/>
+    </logger>
+
+    <logger name="evg.testt.controller">
+        <level value="debug"/>
+        <appender-ref ref="debugLog"/>
+    </logger>
+
+</log4j:configuration>
